Add 'BI' as an exception to account for Power BI
I'm writing a new guide for the Microsoft Visual in Power BI, and the tests are failing on the word "BI". (Error message is: warning  `BI` may be insensitive, use `Bisexual` instead) 

We're unlikely to use "bi" in any scenario other than in referring to Power BI, so I think it's fine for us to add this as an exception!